Combined Login-SendMail
Combination of the two parts
