+++ PluggableAuth/src/main/java/com/Group25/PluggableAuth/EmailLogin/EmailLogin.java
@@ -1,0 +1,12 @@
+package com.Group25.PluggableAuth.EmailLogin;
+
+public class EmailLogin {
+    private String email;
+
+    public String getEmail(){
+        return email;
+    }
+    public void setEmail(String email){
+        this.email = email;
+    }
+}
